requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
csscompressor==0.9.5
Django==4.1.4
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-grappelli==3.0.4
django-guardian==2.4.0
django-rosetta==0.9.8
django-silk==5.0.2
djangorestframework==3.14.0
djangorestframework-csv==2.1.1
ghp-import==2.1.0
gprof2dot==2022.7.29
gunicorn==20.1.0
htmlmin2==0.1.13
idna==3.4
Jinja2==3.1.2
jsmin==3.0.1
Markdown==3.3.7
MarkupSafe==2.1.2
mergedeep==1.3.4
mkdocs==1.4.2
mkdocs-glightbox==0.3.3
mkdocs-material==9.1.6
mkdocs-material-extensions==1.1.1
mkdocs-minify-plugin==0.6.4
mkdocs-static-i18n==0.56
numpy==1.23.5
packaging==23.1
pandas==1.5.2
polib==1.1.1
pycodestyle==2.10.0
Pygments==2.15.1
PyJWT==2.6.0
pymdown-extensions==9.11
PySide6==6.4.1
PySide6-Addons==6.4.1
PySide6-Essentials==6.4.1
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
pyyaml_env_tag==0.1
qt-material==2.12
regex==2023.3.23
requests==2.28.2
shiboken6==6.4.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
unicodecsv==0.14.1
urllib3==1.26.15
watchdog==3.0.0
